requirements.txt
autopep8==2.0.2
beautifulsoup4==4.13.3
black==24.1.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
flake8==7.0.0
idna==3.10
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pycodestyle==2.11.0
pyflakes==3.2.0
PyMySQL==1.1.1
python-dotenv==1.1.0
requests==2.32.3
soupsieve==2.6
typing_extensions==4.13.0
urllib3==2.3.0
